refactor(recipe): change nbOfElements type to Int and document interface
Updated the RecipeRepository interface by changing the type of the nbOfElements parameter from Long to Int. Added documentation for the interface and its parameters to improve clarity and understanding.

+++ app/src/main/java/com/android/sample/model/recipe/RecipeRepository.kt
@@ -1,6 +1,14 @@
 package com.android.sample.model.recipe
 
+/** Interface for retrieving recipe data. */
 interface RecipeRepository {
 
-  fun random(nbOfElements: Long, onSuccess: (List<Recipe>) -> Unit, onFailure: (Exception) -> Unit)
+  /**
+   * Fetches a specified number of random recipes.
+   *
+   * @param nbOfElements The number of random recipes to fetch.
+   * @param onSuccess Callback that returns the list of fetched recipes.
+   * @param onFailure Callback that is called when an error occurs.
+   */
+  fun random(nbOfElements: Int, onSuccess: (List<Recipe>) -> Unit, onFailure: (Exception) -> Unit)
 }
